% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshCIF_functions.R
\name{csh_censor}
\alias{csh_censor}
\title{Adjust censoring times.}
\usage{
csh_censor(df, csh, fuw)
}
\arguments{
\item{df}{Data frame containing raw data for event status and event times.}

\item{csh}{List giving the model of each specific cause.}

\item{fuw}{The length of the follow-up window in time units.}
}
\description{
\code{csh_censor} adjusts the censoring times given a follow-up window.
}
\details{
Each specific cause must have its own time and status variables.
The main cause of interest must be first.

Specifically, for each specific cause, the list is required to have the
following elements:
\describe{
   \item{Cause}{The name of the specific cause}.
   \item{Time}{A string containing the name of the variable with the time.}
   \item{Status}{A string containing the name of the variable with the censoring status.}
   \item{vars}{A vector of covariates for the specific cause.}
}
}
