% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshCIF_functions.R
\name{csh_CIF}
\alias{csh_CIF}
\title{Estimate the CIF at a single risk time.}
\usage{
csh_CIF(fit, Covariates, csh, risk_times = NULL, alpha, method = "loglog")
}
\arguments{
\item{fit}{A fitted object.}

\item{Covariates}{A data set of covariates and risk times at which to estimate the hazard.}

\item{csh}{List giving the model of each specific cause.}

\item{risk_times}{A vector of times at which to estimate the CIF.}

\item{alpha}{The level of statistical significance.}

\item{method}{Method to use for confidence intervals.  Must be one of "original", "log", or
"loglog".}
}
\value{
The CIF.
}
\description{
\code{csh_CIF} estimates the CIF.
}
\details{
The covariates must have an entry for each covariate used in the model
    as well as risk time.
}
\examples{
data(bmtcrr)

df <- data.frame(
    RTime = bmtcrr$ftime,
    RStatus = ifelse(bmtcrr$Status == 1, 1, 0),
    CTime = bmtcrr$ftime,
    CStatus = ifelse(bmtcrr$Status == 2, 1, 0),
    Age = bmtcrr$Age,
    Sex = ifelse(bmtcrr$Sex == "F", 1, 0),
    CR1 = ifelse(bmtcrr$Phase == "CR1", 1, 0),
    CR2 = ifelse(bmtcrr$Phase == "CR2", 1, 0),
    CR3 = ifelse(bmtcrr$Phase == "CR3", 1, 0),
    Source = ifelse(bmtcrr$Source == "PB", 1, 0)
)

# Fit the model.
CSH <- list(
    list(
        Cause = "Relapse",
        Time = "RTime",
        Status = "RStatus",
        vars = c("Age", "Sex", "CR1", "CR2", "CR3")
    ),
    list(
        Cause = "Competing",
        Time = "CTime",
        Status = "CStatus",
        vars=c("Age", "Sex")
    )
)

dfc <- csh_censor(df = df, csh = CSH, fuw = 1)
fit <- csh_fit(dfc, csh = CSH)
Covariates <- data.frame(Age = 30, Sex = 0, CR1 = 0, CR2 = 0, CR3 = 0, Source = 0)
csh_CIF(fit, Covariates, csh = CSH, risk_times = 50, alpha=0.05, method = "loglog")
csh_CIF(fit, Covariates, csh = CSH, risk_times = 50, alpha=0.05, method = "log")
csh_CIF(fit, Covariates, csh = CSH, risk_times = 50, alpha=0.05, method = "original")

}
