% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshCIF_functions.R
\name{csh_fit}
\alias{csh_fit}
\title{Fit the cause-specific hazard model.}
\usage{
csh_fit(dfc, csh)
}
\arguments{
\item{dfc}{Data pre-processed to be censored properly.}

\item{csh}{List giving the model of each specific cause.}
}
\value{
Returns a list with two components:  the data used and the
   Cox proportional hazard model fits.
}
\description{
\code{csh_fit} fits the cause-specific hazard model specified.
}
