% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshCIF.R
\docType{package}
\name{cshCIF}
\alias{cshCIF}
\title{cshCIF a package for estimating the cause-specific hazard CIF.}
\description{
The cshCIF package provides a basic set of functions for carrying
out the estimation.
}
\details{
Interface and functionality are subject to change.
}
